refactor: LineTest -> SectionsTest
